media-vault-analyzer/requirements.txt
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2
tensorflow==2.13.0
requests==2.31.0
python-magic==0.4.27
ffmpeg-python==0.2.0
exifread==3.0.0
transformers==4.35.0
torch==2.1.0
torchvision==0.16.0
